+++ crates/rpstate/src/migration/fields.rs
@@ -2,14 +2,14 @@ use crate::{MigrationContext, Result};
 
 pub struct FieldDescriptor {
     pub name: &'static str,
-    pub type_hash: u64,
+    pub type_hash: u32,
     pub type_name: &'static str,
 }
 
 pub trait RpStateFields: Sized {
     const FIELDS: &'static [FieldDescriptor];
     const VERSION: u32;
-    const SCHEMA_HASH: u64;
+    const SCHEMA_HASH: u32;
     const PARENT_PREFIX: &'static str;
     const MIGRATION_DEPS: &'static [&'static str];
