+++ crates/rpstate/src/store/meta.rs
@@ -3,14 +3,14 @@ use serde::{Deserialize, Serialize};
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Default)]
 pub struct PrefixMeta {
     pub version: u32,
-    pub hash: u64,
+    pub hash: u32,
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
 pub struct StoredFieldEntry {
     pub name: String,
     pub type_name: String,
-    pub type_hash: u64,
+    pub type_hash: u32,
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
